refactor op resolve logic
